dom/media/tests/mochitest/test_peerConnection_basicScreenshare.html

<!DOCTYPE HTML>
<html>
<head>
  <script type="application/javascript" src="pc.js"></script>
</head>
<body>
<pre id="test">
<script type="application/javascript">
  createHTML({
    bug: "1039666",
    title: "Basic screenshare-only peer connection"
  });

  var test;
  runNetworkTest(function (options) {
    const isWinXP = navigator.userAgent.includes("Windows NT 5.1");
    if (IsMacOSX10_6orOlder() || isWinXP) {
        ok(true, "Screensharing disabled for OSX10.6 and WinXP");
        SimpleTest.finish();
        return;
    }
    test = new PeerConnectionTest(options);
    var constraints = {
      video: {
         mozMediaSource: "screen",
         mediaSource: "screen"
      },
      fake: false
    };
    test.setMediaConstraints([constraints], []);
    test.run();
  });
</script>
</pre>
</body>
</html>
